+++ src/main/controllers/ApplicationDetailsController.ts
@@ -3,9 +3,9 @@ import { Response } from 'express';
 import { AppRequest } from '../definitions/appRequest';
 import { GenericTseApplicationTypeItem } from '../definitions/complexTypes/genericTseApplicationTypeItem';
 import { ErrorPages, TranslationKeys } from '../definitions/constants';
+import { getApplicationDisplayByCode } from '../helpers/ApplicationHelper';
 import { findSelectedGenericTseApplication } from '../helpers/GenericTseApplicationHelper';
 import { getApplicationContent } from '../helpers/controller/ApplicationDetailsHelper';
-import { getApplicationDisplayByCode } from '../helpers/controller/ContactTribunalHelper';
 
 export default class ApplicationDetailsController {
   public get = async (req: AppRequest, res: Response): Promise<void> => {
