+++ src/main/controllers/CopyToOtherPartyController.ts
@@ -7,9 +7,9 @@ import { CaseWithId } from '../definitions/case';
 import { PageUrls, TranslationKeys } from '../definitions/constants';
 import { FormContent, FormFields } from '../definitions/form';
 import { CopyToOtherPartyRadioFormFields } from '../definitions/radios';
+import { getApplicationDisplayByCode } from '../helpers/ApplicationHelper';
 import { getPageContent } from '../helpers/FormHelper';
 import { getLanguageParam } from '../helpers/RouterHelpers';
-import { getApplicationDisplayByCode } from '../helpers/controller/ContactTribunalHelper';
 import UrlUtils from '../utils/UrlUtils';
 
 export default class CopyToOtherPartyController {
